+++ packages/bitcoin/src/msg.ts
@@ -5,6 +5,7 @@ import {
   MsgEncoding,
   NumberIsh,
 } from '@xdefi-tech/chains-core';
+import { UTXO } from '@xdefi-tech/chains-utxo';
 import BigNumber from 'bignumber.js';
 import accumulative from 'coinselect/accumulative';
 import * as UTXOLib from 'bitcoinjs-lib';
@@ -15,7 +16,7 @@ import type { BitcoinProvider } from './chain.provider';
 export interface MsgBody {
   amount: NumberIsh;
   to: string;
-  memo?: string;
+  memo?: string | Uint8Array;
   from: string;
   gasLimit?: NumberIsh; // ByteFee
   decimals?: number;
@@ -39,7 +40,7 @@ export class ChainMsg extends BaseMsg<MsgBody, any> {
 
   async buildTx() {
     const msgData = this.toData();
-    let utxos = await this.provider.utxoDataSource.scanUTXOs(this.data.from);
+    let utxos: UTXO[] = await this.provider.scanUTXOs(this.data.from);
     // check is the user has ordinals to erase them from utxos
     let ordinals: any[] = [];
     try {
@@ -124,11 +125,21 @@ export class ChainMsg extends BaseMsg<MsgBody, any> {
     };
   }
 
-  public compileMemo(memo: string) {
-    return UTXOLib.script.compile([
-      UTXOLib.opcodes.OP_RETURN,
-      Buffer.from(memo, 'utf8'),
-    ]);
+  /**
+   * Current method in used to compile a bitcoinjs-lib script. Bitcoin scripts are used to define the conditions
+   * under which funds can be spent in a Bitcoin transaction. Mark a transaction output as unspendable
+   * @param {string | Uint8Array} memo
+   * @returns {Buffer} OP_RETURN compiled script
+   */
+  public compileMemo(memo: string | Uint8Array) {
+    let formattedMemo: Buffer;
+    if (typeof memo === 'string') {
+      formattedMemo = Buffer.from(memo, 'utf8');
+    } else {
+      formattedMemo = Buffer.from(memo);
+    }
+
+    return UTXOLib.script.compile([UTXOLib.opcodes.OP_RETURN, formattedMemo]);
   }
 
   async getFee(speed?: GasFeeSpeed): Promise<FeeEstimation> {