+++ packages/cosmos/src/utils.ts
@@ -7,7 +7,6 @@ import { AminoTypes } from '@cosmjs/stargate/build/aminotypes';
 import { toUtf8 } from '@cosmjs/encoding';
 import { MsgExecuteContract } from 'cosmjs-types/cosmwasm/wasm/v1/tx';
 import { Registry } from '@cosmjs/proto-signing';
-import { osmosisProtoRegistry, osmosisAminoConverters } from 'osmojs';
 
 /**
  * The Amino JSON representation of [MsgExecuteContract].
@@ -30,12 +29,10 @@ export interface AminoMsgExecuteContract {
 export const STARGATE_CLIENT_OPTIONS = {
   registry: new Registry([
     ...defaultRegistryTypes,
-    ...osmosisProtoRegistry,
     ['/cosmwasm.wasm.v1.MsgExecuteContract', MsgExecuteContract],
   ]),
   aminoTypes: new AminoTypes({
     ...createDefaultAminoConverters(),
-    ...osmosisAminoConverters,
     '/cosmwasm.wasm.v1.MsgExecuteContract': {
       aminoType: 'wasm/MsgExecuteContract',
       toAmino: ({