+++ packages/dogecoin/src/chain.provider.ts
@@ -1,4 +1,10 @@
-import { Chain, ChainDecorator, MsgEncoding } from '@xdefi-tech/chains-core';
+import {
+  Chain,
+  ChainDecorator,
+  MsgEncoding,
+  Transaction,
+  TransactionData,
+} from '@xdefi-tech/chains-core';
 import { MsgBody, UtxoProvider } from '@xdefi-tech/chains-utxo';
 
 import { IndexerDataSource } from './datasource';
@@ -10,6 +16,8 @@ import { ChainMsg } from './msg';
   features: [Chain.ChainFeatures.TOKENS],
 })
 export class DogecoinProvider extends UtxoProvider {
+  declare dataSource: IndexerDataSource;
+
   createMsg(
     data: MsgBody,
     encoding: MsgEncoding = MsgEncoding.object
@@ -21,5 +29,17 @@ export class DogecoinProvider extends UtxoProvider {
     return {
       IndexerDataSource: IndexerDataSource,
     };
+  }
+
+  async broadcast(messages: ChainMsg[]): Promise<Transaction[]> {
+    return this.dataSource.broadcast(messages);
+  }
+
+  async getTransaction(txHash: string): Promise<TransactionData | null> {
+    return this.dataSource.getTransaction(txHash);
+  }
+
+  public async scanUTXOs(address: string) {
+    return this.dataSource.scanUTXOs(address);
   }
 }