+++ packages/dogecoin/src/datasource/indexer/queries/index.ts
@@ -1,3 +1,6 @@
 export * from './transactions.query';
 export * from './balances.query';
 export * from './fees.query';
+export * from './broadcast.query';
+export * from './scanUTXOs.query';
+export * from './getTransactionByHash.query';